Rename folder for frequency plots to plots
