+++ src/components/plots/FrequencyPlot.jsx
